Fix for AVAHI under Linux
This version adds a patch to #ifdef around error messages that are
present in the Apple version of the dns_sd.h file, but no the version
distributed with Avahi.

Updated to the latest Tcl Extension Architecture macros.

Reran autoconf

Modified the comments in the function declarations within txt_record.c
to make the file readable my makeheaders.